+++ DEVELOPER/CHANGES.md
@@ -57,6 +57,11 @@
 - Lowered time it takes to consume any of the Bodily Functions medicine types
 - Groin (Bladder) starts developing muscle strain after 60% fullness, scaling proportionally.
 
+- Added functions to BF_ClothingConfig to make adding compatibility with clothing mods (using the overlays) easier:
+[  BF_RegisterClothingCategory(name, data)  ]
+[  BF_AddSoilableLocation(location)]
+[  BF_AddClothingTypesToCategory(category, newTypes)  ]
+
 ==============
 [VERSION 0.11.0]
 ==============